# docs/architecture/api/openapi-spec.yaml
openapi: 3.1.0
info:
  title: ForgeOS Orchestration Platform REST API
  version: 1.0.0
  description: |
    REST API for the ForgeOS distributed multi-agent orchestration platform.
    Provides administrative and operator endpoints for ticket management,
    pipeline overview, health monitoring, and real-time updates.

    This API complements the MCP JSON-RPC protocol used by agent clients.
    While agents interact via MCP tools over Streamable HTTP (`/mcp`),
    operators and dashboards consume these REST endpoints for visibility
    and manual intervention.

    ## Authentication

    All endpoints (except `/health`) require authentication via one of:
    - **Bearer token**: Session token obtained from agent login
    - **API key**: Pre-provisioned key passed in `X-API-Key` header

    ## Error Handling

    All error responses follow the `ErrorResponse` schema with a machine-
    readable `error_code` from the `ForgeOSErrorCode` enum, enabling
    programmatic error handling by clients.

    ## Rate Limiting

    Clients exceeding the configured rate limit receive `429 Too Many Requests`
    with a `Retry-After` header.
  contact:
    name: ForgeOS Team
    url: https://github.com/forgeos
    email: support@forgeos.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  x-last-reviewed: '2026-03-07T09:15:00Z'
  x-diataxis-quadrant: reference
  x-audience: Backend Engineers, Frontend Engineers, DevOps Engineers, Dashboard Developers
  x-related-docs:
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Session token obtained from agent authentication.
        Pass as `Authorization: Bearer <token>`.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Pre-provisioned API key for agent authentication.
        Pass as `X-API-Key: <key>`.

  # ── Reusable Parameters ──
  parameters:
    TicketId:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: ^[A-Z0-9\-]+$
        examples:
          - FORGEOS-ARCH001
          - TASK-FOS-02-002
      description: Human-readable ticket identifier (e.g., `FORGEOS-ARCH001`)
    StageFilter:
      name: stage
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/TicketStage'
      description: Filter tickets by SDLC pipeline stage
    TypeFilter:
      name: type
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/TicketType'
      description: Filter tickets by classification type
    PriorityFilter:
      name: priority
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/TicketPriority'
      description: Filter tickets by priority level
    StatusFilter:
      name: status
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/TicketStatus'
      description: Filter tickets by operational status
    ClaimedByFilter:
      name: claimed_by
      in: query
      required: false
      schema:
        type: string
      description: Filter tickets by claiming agent name
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-indexed)
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page

  # ── Schemas ──
  schemas:
...
    TicketStatus:
      type: string
      enum:
        - READY
        - BLOCKED
        - CLAIMED
        - IN_PROGRESS
        - DONE
        - FAILED
        - ESCALATED
      description: |
        Ticket lifecycle status:
        - `READY` — Unblocked, available for claim
        - `BLOCKED` — Waiting on unresolved dependencies
        - `CLAIMED` — Locked by an agent with active lease
        - `IN_PROGRESS` — Agent actively working
        - `DONE` — Completed successfully
        - `FAILED` — Terminal failure
        - `ESCALATED` — Rework limit exceeded; requires human intervention
    TicketStage:
      type: string
      enum:
        - READY
        - RESEARCH
        - ARCHITECT
        - PRODUCT_MANAGER
        - UI_DESIGN
        - BACKEND
        - FRONTEND
        - QA
        - SECURITY
        - CI
        - DOCUMENTATION
        - VALIDATOR
        - DONE
      description: |
        SDLC pipeline stage. Each ticket type traverses a defined subset
        in strict order. Maps to specialized agents:
        - `READY` — Entry point; ticket is unblocked
        - `RESEARCH` — Research Analyst performs evidence-based analysis
        - `ARCHITECT` — Architect produces ADRs and API contracts
        - `PRODUCT_MANAGER` — Product Manager defines requirements
        - `UI_DESIGN` — UIDesigner creates mockups and design specs
        - `BACKEND` — Backend Engineer implements server-side logic
        - `FRONTEND` — Frontend Engineer implements UI components
        - `QA` — QA Engineer verifies test coverage and correctness
        - `SECURITY` — Security Engineer runs STRIDE/OWASP review
        - `CI` — CI Reviewer checks lint, types, and complexity
        - `DOCUMENTATION` — Documentation Specialist writes docs
        - `VALIDATOR` — Validator performs independent DoD review
        - `DONE` — Terminal state; lifecycle complete
    TicketType:
      type: string
      enum:
        - backend
        - frontend
        - fullstack
        - infra
        - security
        - docs
        - research
        - architecture
        - product
        - design
      description: |
        Ticket classification type. Determines which SDLC flow
        (stage sequence) the ticket follows.
    TicketPriority:
      type: string
      enum:
        - critical
        - high
        - medium
        - low
      description: |
        Ticket priority level. Higher-priority tickets are claimed first.
        - `critical` — Blocking other work; process immediately
        - `high` — Important; process before medium/low
        - `medium` — Standard priority (default)
        - `low` — Process when no higher-priority work is available
...
    ForgeOSErrorCode:
      type: string
      enum:
        - TICKET_NOT_FOUND
        - ALREADY_CLAIMED
        - NOT_CLAIM_OWNER
        - FILE_CONFLICT
        - INVALID_TRANSITION
        - MISSING_EVIDENCE
        - INVALID_SUBTASK
        - LEASE_EXPIRED
        - LEASE_TOO_LONG
        - RATE_LIMITED
        - UNAUTHORIZED
        - FORBIDDEN
        - INTERNAL_ERROR
        - DB_UNAVAILABLE
      description: |
        Machine-readable error code for programmatic error handling.
        - `TICKET_NOT_FOUND` — Requested ticket does not exist
        - `ALREADY_CLAIMED` — Ticket is claimed by another agent
        - `NOT_CLAIM_OWNER` — Caller does not own the claim
        - `FILE_CONFLICT` — File is locked by another ticket
        - `INVALID_TRANSITION` — Stage transition violates SDLC flow
        - `MISSING_EVIDENCE` — Completion attempted without required evidence
        - `INVALID_SUBTASK` — Spawned subtask violates parent constraints
        - `LEASE_EXPIRED` — Agent's lease has expired
        - `LEASE_TOO_LONG` — Requested lease exceeds project maximum
        - `RATE_LIMITED` — API rate limit exceeded
        - `UNAUTHORIZED` — Missing or invalid credentials
        - `FORBIDDEN` — Insufficient permissions
        - `INTERNAL_ERROR` — Unexpected server error
        - `DB_UNAVAILABLE` — PostgreSQL database unreachable

    # ── Domain Model Schemas ──
    Ticket:
      type: object
      description: |
        Ticket entity representing a unit of work in the ForgeOS orchestration
        system. Maps 1:1 to the `tickets` PostgreSQL table (28 fields).
      required:
        - id
        - ticket_id
        - title
        - type
        - priority
        - status
        - stage
        - sdlc_flow
        - depends_on
        - file_paths
        - acceptance_criteria
        - tags
        - rework_count
        - max_reworks
        - lease_duration_minutes
        - metadata
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: UUID primary key (database-generated)
          examples:
            - 550e8400-e29b-41d4-a716-446655440000
        ticket_id:
          type: string
          description: Human-readable ticket identifier (unique)
          examples:
            - FORGEOS-ARCH001
            - TASK-FOS-02-002
        project_id:
          type: string
          format: uuid
          nullable: true
          description: FK to the projects table. Null for unscoped tickets.
        title:
          type: string
          description: Short summary of the work to be done
          examples:
            - Design REST API OpenAPI Specification
        description:
          type: string
          nullable: true
          description: Detailed description of requirements and context
        type:
          $ref: '#/components/schemas/TicketType'
        priority:
          $ref: '#/components/schemas/TicketPriority'
        status:
          $ref: '#/components/schemas/TicketStatus'
        stage:
          $ref: '#/components/schemas/TicketStage'
        sdlc_flow:
          type: array
          items:
            $ref: '#/components/schemas/TicketStage'
          description: |
            Ordered sequence of stages this ticket must traverse.
            Derived from the ticket type's SDLC flow definition.
          examples:
            - - READY
              - ARCHITECT
              - DOCUMENTATION
              - VALIDATOR
              - DONE
        claimed_by:
          type: string
          format: uuid
          nullable: true
          description: UUID of the agent holding the claim. Null if unclaimed.
        claimed_by_name:
          type: string
          nullable: true
          description: Human-readable name of the claiming agent
          examples:
            - Backend Engineer
        machine_id:
          type: string
          nullable: true
          description: Hostname of the machine where the claiming agent runs
          examples:
            - forgeos-dev
        operator:
          type: string
          nullable: true
          description: Human operator who initiated the agent
          examples:
            - Ticketer
        lease_expiry:
          type: string
          format: date-time
          nullable: true
          description: ISO 8601 timestamp when the current lease expires. Null if unclaimed.
        lease_duration_minutes:
          type: integer
          minimum: 1
          default: 30
          description: Default lease duration in minutes (configurable per project)
        depends_on:
          type: array
          items:
            type: string
          description: Array of ticket_id values that must reach DONE before this ticket unblocks
          examples:
            - - FORGEOS-ARCH001
              - FORGEOS-RES005
        file_paths:
          type: array
          items:
            type: string
          description: Workspace-relative file paths within this ticket's write scope
          examples:
            - - docs/architecture/api/openapi-spec.yaml
        acceptance_criteria:
          type: array
          items:
            type: string
          description: Conditions that must be satisfied for the ticket to pass validation
        tags:
          type: array
          items:
            type: string
          description: Free-form labels for filtering and categorization
          examples:
            - - architecture
              - api
              - phase1
        rework_count:
          type: integer
          minimum: 0
          description: Number of times this ticket has been sent back to rework
        max_reworks:
          type: integer
          minimum: 1
          default: 3
          description: Maximum allowed rework attempts before escalation
        metadata:
          type: object
          additionalProperties: true
          description: |
            Arbitrary JSONB metadata (e.g., evidence, agent notes, stage artifacts).
            Merged with stage evidence on advance.
          examples:
            - architecture_adr: ADR-005
              coverage: 85
        parent_id:
          type: string
          nullable: true
          description: ticket_id of the parent ticket if this is a spawned subtask
        source_task_file:
          type: string
          nullable: true
          description: Path to the TODO task file that generated this ticket
          examples:
            - TODO/tasks/phase1-architecture.md
        created_at:
          type: string
          format: date-time
          description: ISO 8601 creation timestamp
        updated_at:
          type: string
          format: date-time
          description: ISO 8601 last-modified timestamp
        completed_at:
          type: string
          format: date-time
          nullable: true
          description: ISO 8601 completion timestamp. Null until ticket reaches DONE.
    TicketListResponse:
      type: object
      description: Paginated list of tickets
      required:
        - tickets
        - total
        - page
        - page_size
        - total_pages
      properties:
        tickets:
          type: array
          items:
            $ref: '#/components/schemas/Ticket'
        total:
          type: integer
          minimum: 0
          description: Total number of tickets matching the filter criteria
        page:
          type: integer
          minimum: 1
          description: Current page number
        page_size:
          type: integer
          minimum: 1
          description: Number of items per page
        total_pages:
          type: integer
          minimum: 0
          description: Total number of pages

    # ── Event Schemas ──
...
    StageInfo:
      type: object
      description: Information about a single SDLC pipeline stage
      required:
        - name
        - description
        - agent
        - ticket_count
      properties:
        name:
          $ref: '#/components/schemas/TicketStage'
        description:
          type: string
          description: Human-readable description of the stage's purpose
          examples:
            - "Backend Engineer implements server-side logic"
        agent:
          type: string
          description: The agent role responsible for this stage
          examples:
            - Backend Engineer
        ticket_count:
          type: integer
          minimum: 0
          description: Number of tickets currently in this stage
    StageOverviewResponse:
      type: object
      description: Pipeline stage overview with counts and SDLC flow definitions
      required:
        - stages
        - sdlc_flows
        - total_tickets
      properties:
        stages:
          type: array
          items:
            $ref: '#/components/schemas/StageInfo'
          description: All pipeline stages with current ticket counts
        sdlc_flows:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/TicketStage'
          description: |
            SDLC flow definitions per ticket type. Maps each TicketType
            to its ordered stage sequence.
          examples:
            - backend:
                - READY
                - BACKEND
                - QA
                - SECURITY
                - CI
                - DOCUMENTATION
                - VALIDATOR
                - DONE
              architecture:
                - READY
                - ARCHITECT
                - DOCUMENTATION
                - VALIDATOR
                - DONE
        total_tickets:
          type: integer
          minimum: 0
          description: Total number of tickets across all stages

    # ── Health Schemas ──
...
    ErrorResponse:
      type: object
      description: |
        Structured error response returned by all endpoints on failure.
        Clients can switch on `error_code` for programmatic handling and
        display `message` for human-readable context.
      required:
        - status
        - error_code
        - message
        - timestamp
      properties:
        status:
          type: integer
          description: HTTP status code
          examples:
            - 404
            - 409
            - 401
        error_code:
          $ref: '#/components/schemas/ForgeOSErrorCode'
        message:
          type: string
          description: Human-readable error description
          examples:
            - "Ticket FORGEOS-ARCH001 not found"
        details:
          type: object
          additionalProperties: true
          nullable: true
          description: Optional additional context (e.g., conflicting file paths)
          examples:
            - conflicting_files:
                - src/server.ts
              locked_by: TASK-FOS-02-002
        ticket_id:
          type: string
          nullable: true
          description: The ticket involved in the error, if applicable
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the error occurred

    # ── WebSocket Schemas ──
...
paths:
...
  /tickets:
    get:
      summary: List tickets
      description: |
        Retrieve a paginated list of tickets with optional filters.
        Supports filtering by stage, type, priority, status, and claiming agent.
        Results are ordered by priority (critical first), then by creation date.
      operationId: listTickets
      tags:
        - Tickets
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/StageFilter'
        - $ref: '#/components/parameters/TypeFilter'
        - $ref: '#/components/parameters/PriorityFilter'
        - $ref: '#/components/parameters/StatusFilter'
        - $ref: '#/components/parameters/ClaimedByFilter'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Paginated list of tickets matching the filter criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketListResponse'
              examples:
                default:
                  summary: Tickets in BACKEND stage
                  value:
                    tickets:
                      - id: 550e8400-e29b-41d4-a716-446655440000
                        ticket_id: TASK-FOS-02-002
                        project_id: null
                        title: Implement MCP Server Core
                        description: null
                        type: backend
                        priority: high
                        status: CLAIMED
                        stage: BACKEND
                        sdlc_flow:
                          - READY
                          - BACKEND
                          - QA
                          - SECURITY
                          - CI
                          - DOCUMENTATION
                          - VALIDATOR
                          - DONE
                        claimed_by: 660e8400-e29b-41d4-a716-446655440001
                        claimed_by_name: Backend Engineer
                        machine_id: forgeos-dev
                        operator: Ticketer
                        lease_expiry: "2026-03-07T08:30:00Z"
                        lease_duration_minutes: 30
                        depends_on:
                          - FORGEOS-ARCH001
                        file_paths:
                          - forgeos-server/src/server.ts
                        acceptance_criteria:
                          - MCP server starts and accepts connections
                        tags:
                          - backend
                          - phase2
                        rework_count: 0
                        max_reworks: 3
                        metadata: {}
                        parent_id: null
                        source_task_file: TODO/tasks/phase2-mcp-server.md
                        created_at: "2026-03-05T17:51:09Z"
                        updated_at: "2026-03-06T12:00:00Z"
                        completed_at: null
                    total: 1
                    page: 1
                    page_size: 20
                    total_pages: 1
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                unauthorized:
                  value:
                    status: 401
                    error_code: UNAUTHORIZED
                    message: "Missing or invalid authentication credentials"
                    details: null
                    timestamp: "2026-03-07T08:00:00Z"
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until the rate limit resets

  # ── Get Ticket ──
  /tickets/{id}:
    get:
      summary: Get ticket by ID
      description: |
        Retrieve a single ticket by its human-readable ticket_id.
        Returns the full ticket object including all metadata.
      operationId: getTicket
      tags:
        - Tickets
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: Ticket details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_found:
                  value:
                    status: 404
                    error_code: TICKET_NOT_FOUND
                    message: "Ticket FORGEOS-NONEXIST not found"
                    ticket_id: FORGEOS-NONEXIST
                    timestamp: "2026-03-07T08:00:00Z"

  # ── Claim Ticket ──
...
  /stages:
    get:
      summary: Pipeline stage overview
      description: |
        Retrieve the list of all SDLC pipeline stages with current ticket
        counts and the SDLC flow definitions per ticket type.

        This endpoint provides the data needed for the dashboard's
        pipeline visualization — a Kanban-style view of tickets moving
        through stages.
      operationId: listStages
      tags:
        - Stages
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        '200':
          description: Pipeline stage overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StageOverviewResponse'
              examples:
                default:
                  value:
                    stages:
                      - name: READY
                        description: Unblocked tickets available for claim
                        agent: System
                        ticket_count: 5
                      - name: BACKEND
                        description: Server-side implementation in progress
                        agent: Backend Engineer
                        ticket_count: 2
                      - name: QA
                        description: Test verification in progress
                        agent: QA Engineer
                        ticket_count: 1
                      - name: DONE
                        description: Lifecycle complete
                        agent: System
                        ticket_count: 12
                    sdlc_flows:
                      backend:
                        - READY
                        - BACKEND
                        - QA
                        - SECURITY
                        - CI
                        - DOCUMENTATION
                        - VALIDATOR
                        - DONE
                      frontend:
                        - READY
                        - UI_DESIGN
                        - FRONTEND
                        - QA
                        - SECURITY
                        - CI
                        - DOCUMENTATION
                        - VALIDATOR
                        - DONE
                    total_tickets: 20
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Health Check ──
...